+++ src/EntwineLLM/Commands/Interfaces/IBaseCommand.cs
@@ -4,7 +4,7 @@ namespace EntwineLlm.Commands.Interfaces
 {
     internal interface IBaseCommand
     {
-        int Id {  get; }
+        int Id { get; }
         void Execute(object sender, EventArgs e);
     }
 }