+++ src/EntwineLLM/EntwineLLMOptions.cs
@@ -4,7 +4,7 @@ using System.ComponentModel;
 
 namespace EntwineLlm.Models
 {
-    public class EntwineLlmOptions : DialogPage
+    public class GeneralOptions : DialogPage
     {
         [Category("Configuration")]
         [DisplayName("Large Language Model Base Url")]
@@ -16,10 +16,33 @@ namespace EntwineLlm.Models
         [Description("Sets timeout for HTTP requests")]
 
         public TimeSpan LlmRequestTimeOut { get; set; } = new TimeSpan(0, 10, 0);
+    }
 
-        [Category("Configuration")]
-        [DisplayName("LLM Model")]
-        [Description("Sets the model to be used when querying LLM")]
-        public string LlmModel { get; set; } = "llama3.2";
+    public class ModelsOptions : DialogPage
+    {
+        [Category("Models")]
+        [DisplayName("Refactor queries")]
+        [Description("Sets the model to be used when querying LLM for refactor")]
+        public string LlmRefactor { get; set; } = "llama3.2";
+
+        [Category("Models")]
+        [DisplayName("Unit tests generation")]
+        [Description("Sets the model to be used when querying LLM for unit tests generation")]
+        public string LlmUnitTests { get; set; } = "llama3.2";
+
+        [Category("Models")]
+        [DisplayName("Documentation generation")]
+        [Description("Sets the model to be used when querying LLM for code documentation generation")]
+        public string LlmDocumentation { get; set; } = "llama3.2";
+
+        [Category("Models")]
+        [DisplayName("Code review query")]
+        [Description("Sets the model to be used when querying LLM for code review")]
+        public string LlmReview { get; set; } = "llama3.2";
+
+        [Category("Models")]
+        [DisplayName("Follow-up query")]
+        [Description("Sets the model to be used when querying LLM for follow-up prompts")]
+        public string LlmFollowUp { get; set; } = "llama3.2";
     }
 }