+++ src/EntwineLLM/Helpers/ProjectHelper.cs
@@ -14,16 +14,8 @@ namespace EntwineLlm.Helpers
             var dte = Package.GetGlobalService(typeof(DTE)) as DTE2
                 ?? throw new InvalidOperationException("DTE not available");
 
-            var activeProject = GetActiveProject(dte);
-
-            if (activeProject != null)
-            {
-                activeProject.ProjectItems.AddFromFileCopy(filePath);
-            }
-            else
-            {
-                throw new InvalidOperationException("No active project found.");
-            }
+            var activeProject = GetActiveProject(dte) ?? throw new InvalidOperationException("No active project found.");
+            activeProject.ProjectItems.AddFromFileCopy(filePath);
         }
 
         public static Project GetActiveProject(DTE2 dte)
@@ -31,12 +23,7 @@ namespace EntwineLlm.Helpers
             ThreadHelper.ThrowIfNotOnUIThread();
 
             Array activeSolutionProjects = (Array)dte.ActiveSolutionProjects;
-            if (activeSolutionProjects.Length > 0)
-            {
-                return activeSolutionProjects.GetValue(0) as Project;
-            }
-
-            return null;
+            return activeSolutionProjects.Length == 0 ? null : activeSolutionProjects.GetValue(0) as Project;
         }
     }
 }